+++ src/main/java/com/itilria/altograce/repository/ConductRepository.java
@@ -5,6 +5,6 @@ import com.itilria.altograce.domain.CodeOfConduct;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ConductRepository extends JpaRepository<CodeOfConduct, Integer>{
+public interface ConductRepository extends JpaRepository<CodeOfConduct, Long>{
 
 }
