+++ src/main/java/com/itilria/altograce/repository/clientrepository/ClientDependencyRepository.java
@@ -8,8 +8,8 @@ import org.springframework.data.jpa.repository.Query;
 
 import com.itilria.altograce.domain.client.ClientDependency;
 
-public interface ClientDependencyRepository extends JpaRepository<ClientDependency, Integer>{
-    List<ClientDependency> findByPrimaryClient_Id(int id);
+public interface ClientDependencyRepository extends JpaRepository<ClientDependency, Long>{
+    List<ClientDependency> findByPrimaryClient_Id(long id);
     @Query("SELECT u FROM ClientDependency u WHERE u.id_passport = ?1")
     Optional<ClientDependency> findByPassport(String id_passport);
 }
