+++ src/main/java/com/itilria/altograce/repository/clientrepository/DeceasedRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.itilria.altograce.domain.client.Deceased;
 
-public interface DeceasedRepository extends JpaRepository<Deceased, Integer>{
-    List<Deceased> findByPrimaryClient_Id(int id);
+public interface DeceasedRepository extends JpaRepository<Deceased, Long>{
+    List<Deceased> findByPrimaryClient_Id(long id);
     Optional<Deceased> findByBiNumberAndGraveNumber(String biNumber, String graveNumber);
 }
