+++ src/main/java/com/itilria/altograce/repository/clientrepository/SecondaryPackageRepository.java
@@ -5,7 +5,7 @@ import com.itilria.altograce.domain.client.SecondaryPackageSubscription;
 import java.util.Optional;
 import java.util.List;
 
-public interface SecondaryPackageRepository extends JpaRepository<SecondaryPackageSubscription, Integer>{
+public interface SecondaryPackageRepository extends JpaRepository<SecondaryPackageSubscription, Long>{
     //Optional<PrimaryPackageSubscription> findById(int id);
     //List<PrimaryPackageSubscription> findByPrimaryClient_Clientid(String clientId);
 }
