+++ src/main/java/com/itilria/altograce/service/ConductService.java
@@ -18,7 +18,7 @@ public class ConductService{
     @Autowired
     private CompanyRepository companyRepository;
 
-    public CodeOfConduct findById(int id)
+    public CodeOfConduct findById(long id)
     {
         return repository.findById(id).orElse(null);
     }
